requirements.txt
numpy==2.3.2
pillow==11.3.0
streamlit==1.48.1
tensorflow==2.20.0
